Move SnagNavDisplay + InlineDialogSceneStrategy to design/fe (#231)
The inline strategy and SnagNavDisplay both belong with the other
navigation/scene UI primitives in lib/design/fe rather than in
lib/routing/fe, which stays the low-level holder for SnagBackStack
and SnagNavRoute types.

Dependency direction: design/fe now depends on routing/fe (types
only); routing/fe no longer needs design/fe.
