+++ composeApp/src/commonMain/kotlin/cz/adamec/timotej/snag/ui/MainNavigation.kt
@@ -13,8 +13,8 @@
 package cz.adamec.timotej.snag.ui
 
 import androidx.compose.runtime.Composable
+import cz.adamec.timotej.snag.lib.design.fe.navigation.SnagNavDisplay
 import cz.adamec.timotej.snag.lib.design.fe.scenes.ContentPaneSceneStrategy
-import cz.adamec.timotej.snag.lib.navigation.fe.SnagNavDisplay
 import org.koin.compose.koinInject
 
 @Composable
