+++ feat/projects/fe/driving/api/src/commonMain/kotlin/cz/adamec/timotej/snag/projects/fe/driving/api/ProjectsNavigation.kt
@@ -14,8 +14,8 @@ package cz.adamec.timotej.snag.projects.fe.driving.api
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import cz.adamec.timotej.snag.lib.design.fe.navigation.SnagNavDisplay
 import cz.adamec.timotej.snag.lib.design.fe.scenes.ContentPaneSceneStrategy
-import cz.adamec.timotej.snag.lib.navigation.fe.SnagNavDisplay
 import org.koin.compose.koinInject
 
 @Composable
