+++ lib/design/fe/src/commonMain/kotlin/cz/adamec/timotej/snag/lib/design/fe/scenes/InlineDialogSceneStrategy.kt
@@ -12,7 +12,7 @@
 
 @file:Suppress("UnnecessaryFullyQualifiedName")
 
-package cz.adamec.timotej.snag.lib.navigation.fe
+package cz.adamec.timotej.snag.lib.design.fe.scenes
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
